+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 0.0.1 - Initial version.
 0.0.2 - update - use @opensolar/ossdk-react package, update links to other sdk sites
+
+0.0.3 - - Various minor bug fixes
+- Updated package versions
+- Added setbacks button
+- Improved formatting of data output